chore[reorganize-files]: reorganizing folders and files and [WIP] started to separate javascript code between files
