+++ src/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <title>Jiji Movies</title>
-    <link rel="stylesheet" href="styles/home.css" />
+    <link rel="stylesheet" href="./styles/pages/home.css" />
     <link rel="icon" href="assets/logo.svg" alt="logo" />
   </head>
   <body>
@@ -40,6 +40,6 @@
         Created by Karina Mel and Ludmila Magalhães
       </footer>
     </div>
-    <script src="/src/js/main.js" type="module"></script>
+    <script src="./js/pages/home.js" type="module"></script>
   </body>
 </html>
