+++ src/js/components/header.js
@@ -1,0 +1,8 @@
+let toggleButton = document.getElementById("toggleButton");
+let navList = document.getElementById("navList");
+
+export function menu() {
+  toggleButton.addEventListener("click", () => {
+    navList.classList.toggle("active");
+  });
+}
