+++ src/movies.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="utf-8" />
     <title>Jiji Movies</title>
-    <link rel="stylesheet" href="styles/movies.css" />
+    <link rel="stylesheet" href="./styles/pages/movies.css" />
     <link rel="icon" href="assets/logo.svg" alt="logo" />
   </head>
   <body>
@@ -47,6 +47,6 @@
         Created by Karina Mel and Ludmila Magalhães
       </footer>
     </div>
-    <script src="js/main.js" type="module"></script>
+    <script src="./js/pages/movies.js" type="module"></script>
   </body>
 </html>
