--- src/styles/characters.css
@@ -1,3 +1,0 @@
-@import "/src/styles/reset.css";
-@import "/src/styles/header.css";
-@import "/src/styles/footer.css";
