+++ src/styles/components/footer.css
@@ -1,4 +1,4 @@
-@import "variables.css";
+@import "../variables.css";
 
 .footerContainer {
   background-color: #015249;
