+++ src/styles/components/header.css
@@ -1,4 +1,4 @@
-@import "variables.css";
+@import "../variables.css";
 
 .navBar {
   background-color: #ff3b3f;
