+++ src/styles/components/input-search.css
@@ -1,0 +1,1 @@
+@import "../variables.css";
