+++ src/styles/components/input-select.css
@@ -1,0 +1,1 @@
+@import "../variables.css";
