+++ src/styles/pages/characters.css
@@ -1,0 +1,3 @@
+@import "/src/styles/reset.css";
+@import "/src/styles/components/header.css";
+@import "/src/styles/components/footer.css";
